[analyzer] Improve dead code highlighting for null-aware expressions
Previously, when a null-aware property access, method invocation, or
index invocation was performed on a variable that was definitely
unassigned, the entire expression was highlighted as dead code.

This change modifies the `DeadCodeVerifier` to highlight only from the
null-aware operator (`?.`, `?[]`, etc.) to the end of the expression.
This provides a more precise and helpful diagnostic by highlighting
the exact portion of the code that can be safely removed.

Fixes https://github.com/dart-lang/sdk/issues/61251.

Bug: https://github.com/dart-lang/sdk/issues/61251
Change-Id: I4792b84d16a797f70b752a20e6affa69f3fe3d8c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444040
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>