[analyzer/linter] Clean up error message comments.
This change modifies the comments in the analyzer and linter
`messages.yaml` files so that:

- Parameters are always described at the end of the comment. If the
  message takes no parameters, the comment ends with the text `No
  parameters.`

- Parameter descriptions are always word wrapped to 80 columns.

It also fixes the format of the comment field for
`ANALYSIS_OPTION_DEPRECATED` to remove an unintentional blank line,
and adds documentation for the second parameter of
`INVALID_EXPORT_OF_INTERNAL_ELEMENT_INDIRECTLY` (which was previously
missing).

In a follow-up CL, I will transition to a more structured format for
parameters in these `messages.yaml` files, and the `Parameters:`
sections of the comments will be automatically generated from this
structured format during code generation. Cleaning up the comments now
will ensure that the follow-up CL is clean (makes no changes to
generated files), so we'll be able to have high confidence in its
correctness.

Change-Id: I6aa1f75fa0b8e4fb9724dcba00fb7d1f06dfb355
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444205
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>