Fine. Update LinkedElementFactory.libraryOfUri() to return null for missing library.
We use this method to check requirements, and it is possible that some
libraries are not available anymore. Specifically, this happens when
conditional imports used, and environment variables changed.

There are no new tests, because test failures happened with existing
summary tests when `withFineDependencies` is `true`.

Change-Id: I460948a759be7ed8c5b36f76b61ab4e1d1522097
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444520
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>