[VM/Service] Tweak profile_startup_cli_flag_test to make it less flaky
This is a followup to
https://github.com/dart-lang/sdk/commit/1ddbc4428829f126a7963ca11763bec90187fe5f
that follows the same rationale described in that commit’s message. This
followup is being made because the original change did not successfully
resolve the flakiness on Windows configurations. If the flakiness is
still present after this change lands, we probably have to replace this
`package:vm_service` integration test with a C++ test that interacts
directly with a `SampleBlockBuffer`.

Issue: https://github.com/dart-lang/sdk/issues/61191
Change-Id: Ifc3029dbd23b3bcfaac947d3b3f6bd54b973eb05
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444443
Reviewed-by: Ben Konyi <bkonyi@google.com>
Commit-Queue: Derek Xu <derekx@google.com>