[analyzer/linter] Document message parameter names and types.
The code generation for diagnostic codes is changed so that parameter
names and their types show up in the auto-generated doc comments. When
we begin the transition to reporting diagnostics using a
statically-checked API, these comments should make it easier to figure
out what parameters to pass in.

Change-Id: I6b33ccd5b5d5a24001da2e1af2cd31cc30366e6a
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444365
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>