Generate getters and setters for modifiers in fragments.
I noticed that modifiers were always getter/setter pairs and had names
that were derived from the modifier names. Because of this, the entire
code could be reduced to just the modifier's name.

For now I only annotated and generated for `ClassFragmentImpl`,
`ConstructorFragmentImpl`, and `VariableFragmentImpl`. WDYT?

If this looks good, I will follow with a CL to generate all modifiers
for all fragments.

Change-Id: I02d4243706afb49f0cab15198816f6d0e33e165b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/443461
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>