[analyzer] fix CHANGELOG after versioning mistake.
https://dart-review.googlesource.com/c/sdk/+/444461 deprecated
`Diagnostic.data` and added some return types to `DiagnosticReporter`
to give clients an alternative to using it. Hence, it should have been
accompanied by a minor version bump. When I sent it for review, I
failed to realize that analyzer 8.1.0 had already been released, so I
didn't do the necessary bump, and I added a note to the CHANGELOG in
the wrong place.

This change fixes the CHANGELOG. The version bump is no longer needed
since it was done in
https://dart-review.googlesource.com/c/sdk/+/444863.

Change-Id: I6a6a6964c71f5a000aa51a314f5aaef40f92a63f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444901
Auto-Submit: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>