Fine. Record opaque API usages.
For some getters or methods we don't have good way to track them.
So, we just record them, and the requirements will always fail.
The idea is that for practical purposes we also don't need this.

One example is accessible `firstFragment`, with the counter-argument
that for analysis you (analyzer, linter) should only only use elements.

Anyway, this is necessary for correctness, and we can improve step
by step precision later. FWIW, the current state forces 9 files to be
analyzer, 7 of them because of these "opaque" requirement, vs. 1427
without fine-grained dependencies, when analyzing analyzer/ itself.

Change-Id: I6277ae6a7d1760765fdc758a178e520451510fbd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444704
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>