[analysis_server] Dot shorthands: Add ConvertToDotShorthand assist.
An assist to convert `E.id` to its dot shorthand equivalent `.id`. The assist handles constructors, static methods, and static fields/getters.

I kept this as an assist because there's no diagnostic that corresponds to this. If we have a lint that's more opinionated, happy to associate this fix with it, but until then, it's a general assist.

Added a bunch of tests for what I could think of testing. Let me know if there are any suggestions.

Bug: https://github.com/dart-lang/sdk/issues/60994
Change-Id: I30b9670103e516bccf8aa46b704c353910796629
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444202
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>