[analyzer] Prepare to rename error codes to camelCase.
A temporary script is added,
`pkg/analyzer/tool/messages/rename_error_constants.dart`, which will
rename every error code in the analyzer (and related packages) from
SCREAMING_CAPS to camelCase.

Also, the error message code generator is enhanced so that it can
generate error codes using either SCREAMING_CAPS or camelCase, based
on the setting of a temporary constant, `_useLowerCamelCaseNames`. The
constant is currently set to `false`; `rename_error_constants.dart`
will change it to `true`.

There are a few references to the old SCREAMING_CAPS names that won't
be able to be automatically addressed by the script. To avoid
breakages, a mechanism has been added to the error message code
generator to preserve those names after the migration, as deprecated
aliases. The names that will be preserved are listed in the
`ErrorClassInfo.deprecatedSnakeCaseNames` field, in
`error_code_info.dart`. These will be cleaned up in future CLs.

Once the rename is complete, the temporary script and constant will be
removed.

Change-Id: I6a6a69640b077c6bc4da92cec97686e4b2c3b906
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/444920
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Moritz Sümmermann <mosum@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>