[analyzer] Fix output generation upon test failure to use camelCase.
When `GatheringDiagnosticListener.assertErrors` detects a test
failure, it outputs a suggestion for how to change the test
expectations to match the current behavior.

Since error code constants are in `camelCase` now, but
`DiagnosticCode.uniqueName` still takes the form
`ClassName.SCREAMING_CAPS` or `ClassName.snake_case`, this logic needs
to be updated to convert the text after the `.` to `camelCase`.

Change-Id: I6a6a696478e0443aab5184ac3331af9aac29a9aa
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445407
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>