Elements. Deprecate supertype, mixins, and interfaces on InterfaceFragment
Deprecate the `supertype`, `mixins`, and `interfaces` getters on
`InterfaceFragment`. This information more accurately belongs on the
`InterfaceElement`, which represents the fully merged view of a class,
rather than on a single fragment declaration.

This change is part of an ongoing effort to consolidate class hierarchy
information at the element level. All internal call sites have been
updated to use the corresponding getters on `InterfaceElement`.

As part of this refactoring, the `isValidMixin` logic has also been
moved from `ClassFragmentImpl` to `ClassElementImpl` to align with this
principle.

Change-Id: Ia919eaa6328451d186e9f67e2c1bb8605856f109
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445561
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>