Elements. Simplify `_writeElementList` test utility
The `_writeElementList` helper in the element text test utility included
complex logic to handle element types that lack a standard
`enclosingElement`. This special casing for `LibraryImport`,
`LibraryExport` (which are not elements at all now), and `PrefixElement`
made the function fragile.

This change removes the conditional logic and tightens the function's
signature to only accept `ElementImpl` (so it cannot be
`SubstitutedXyz`) subtypes. As a result, the call site for `prefixes` is
updated to use the more general `_writeList` function, leaving
`_writeElementList` with the single responsibility of verifying the
`enclosingElement` for applicable types.

Change-Id: I436a863a59bb0394556002bd720c363eacea826b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445620
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>