Elements. Move extension type properties from fragment to element
The `typeErasure`, `primaryConstructor`, and `representation` fields are
properties of the extension type as a whole, not of any individual
declaration fragment.

This commit moves `typeErasure` from `ExtensionTypeFragmentImpl` to
`ExtensionTypeElementImpl` to better align the element model with this
concept.

As part of this refactoring, the `primaryConstructor` and
`representation` getters on `ExtensionTypeFragment` are now deprecated.
Clients should access these properties directly from the
`ExtensionTypeElement`, which serves as the single source of truth. All
internal call sites have been updated accordingly.

Change-Id: I359572c6f7fe1d562fadfb60873a6e34aed0e5ee
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445580
Reviewed-by: Paul Berry <paulberry@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>