[linter] Use LinterLintCode as the type of declared lint codes.
This CL changes the diagnostic code generator for lint diagnostic
codes to consistently use `LinterLintCode` as the static type of the
generated constants, instead of the superclass, `LintCode`.

This change makes the code generation more consistent between the
analyzer and the linter, which will pave the way for further
improvements to diagnostic code generation that I have planned in
follow-up CLs.

Change-Id: I6a6a6964f5b533eb722066195d83bd8379dbbd43
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445781
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>