[dart2wasm] Simplify boxed Dart object marker allocation
Skip the `JSSymbol` indirection and use an `externref` for the `Symbol`
object directly.

With the `JSSymbol` the indirection cannot be eliminated by inlining,
because we have to check whether the `JSSymbol` static object is
allocated first, and then access its `externref` field.

When using an `externref` directly we still check whether the static
field is initialized, but if it is we directly access the `extenref`,
without going through `JSSymbol`.

Change-Id: I9feae319c34fc5a3f83c0f5ba3767454a1f94566
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445680
Reviewed-by: Srujan Gaddam <srujzs@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>