[cfe] Make MethodInvocation, PropertyGet, PropertySet, IndexGet and IndexSet null-aware
This embeds the handling of null-aware access into the regular internal nodes. This also avoids some of the lowerings performed during body building.

This removes the now unused NullAwarePropertyGet and NullAwarePropertySet. NullAwareMethodInvocation will be removed latter because it is still used in other lowerings.

Change-Id: I88f3d67919cc9dadf9d16dcef4a474de90780260
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445940
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Johnni Winther <johnniwinther@google.com>