[analyzer] Dot shorthands: Issue 61209 - Constant parameter mismatch error fix.
In the specific case where a dot shorthand is referencing a const constructor invocation of a imported class and the constructor has an optional parameter with a default value, the analyzer currently fails to compute the the constant result from that default value.

With the changes in this CL, we're properly calculating the dependencies we need to do that computation and prevent the `const_constructor_param_type_mismatch` error. We're missing a couple of visitors for the `DotShorthandConstructorInvocation` AST that tell the constant evaluator to evaluate its dependencies. This should work exactly like visiting `InstanceCreationExpression`s, in the same class.

Fixes: https://github.com/dart-lang/sdk/issues/61209
Bug: https://github.com/dart-lang/sdk/issues/59835
Change-Id: I6544e11e19f814b62108a736d1c778f8c4eda0cc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/446181
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>