[stable][ddc] Avoid accidentally invoking a getter
During the hot reload operation existing class static getters were
accidentally being invoked on the new code being updated. This could
cause the hot reload operation to fail and crash the application when
the side effects of the getter throw an exception.

What is the fix: When checking for members that were deleted during a
hot reload this change tests for the presence of the member name
using the built in operations from the native JavaScript Object to list
the names without invoking the members.

Why cherry-pick: Some users are reporting that hot reload in a web
environment will crash during the operation. This bug can prevent them
from using hot reload in their application.

Risk: Low, this code only runs on a hot reload and was already fixed
on the main channel as part of a larger change. This cherry pick
only targets the bug and avoids introducing the other changes.

This fix is already applied on main branch as part of:

Cherry-pick: https://dart-review.googlesource.com/c/sdk/+/440082
Issue: https://github.com/dart-lang/sdk/issues/61341
Change-Id: I8e23a977ee784a88cfa0381dc2a89f4feb1c8229
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/446440
Reviewed-by: Nate Biggs <natebiggs@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>