[analyzer] Re-work handling of null shorting using shared code.
In https://dart-review.googlesource.com/c/sdk/+/399480, a mechanism
was introduced to `_fe_analyzer_shared` to support null-shorting.
Sometime later, in
https://dart-review.googlesource.com/c/sdk/+/427341, the CFE was
changed to use this shared mechanism.

This CL completes the arc of work by changing the analyzer to use the
new shared mechanism.

The public API class `NullShortableExpression`, which was part of the
analyzer's old implementation of null shorting, is no longer used. It
should never have been exposed through the analyzer public API in the
first place, so it's been deprecated; I will update the pending
changes for analyzer 9.0.0 to delete it.

Change-Id: Ic711df6c537c454aa8a99861135e63c8dd277485
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/400021
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>