[analyzer] Switch error code references in messages.yaml to camelCase
This is a follow-up to
https://dart-review.googlesource.com/c/sdk/+/444921, which renamed
analyzer error codes from SCREAMING_CAPS conventions to camelCase
conventions. That CL left a few deprecated error codes in place
because they were referred to in comments in the `messages.yaml` file,
and those references couldn't be fixed in an automated fashion.

This CL cleans up the references in `messages.yaml` to use the new
camelCase constants, and removes the deprecated SCREAMING_CAPS
constants that are no longer needed.

Change-Id: I6a6a696406efa02d73775d1a8e0defa5a0e8e39b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/445105
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Auto-Submit: Paul Berry <paulberry@google.com>